# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OMS API",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/instruments": {
      "get": {
        "tags": [
          "orders"
        ],
        "operationId": "list_instruments_for_trader",
        "parameters": [
          {
            "name": "search",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Matching active instruments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InstrumentSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No credential"
          }
        },
        "security": [
          {
            "basic_auth": []
          },
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InstrumentSummary": {
        "type": "object",
        "required": [
          "id",
          "symbol",
          "name",
          "venue",
          "asset_class",
          "status"
        ],
        "properties": {
          "asset_class": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "venue": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "basic_auth": {
        "type": "http",
        "scheme": "basic",
        "description": "Username = key_id, password = secret."
      },
      "bearer_token": {
        "type": "http",
        "scheme": "bearer",
        "description": "User endpoints: a trading token, `key_id.secret` (Databento-style). Admin endpoints (/admin/*): the static admin token."
      }
    }
  },
...
}